+++ adept/container/base/__init__.py
@@ -1,1 +1,2 @@
 from .container import Container
+from .nccl_optimizer import NCCLOptimizer
