+++ adept/scripts/actorlearner.py
@@ -216,24 +216,15 @@ def main(args):
         done_closing = ray.wait(closes)
 
     if args.eval:
-        import subprocess
-        command = [
-            'python',
-            '-m',
-            'adept.scripts.evaluate',
-            '--log-id-dir',
-            log_id_dir,
-            '--gpu-id',
-            str(args.gpu_id),
-            '--nb-episode',
-            str(30)
-        ]
+        from adept.scripts.evaluate import main
+        eval_args = {
+            'log_id_dir': log_id_dir,
+            'gpu_id': 0,
+            'nb_episode': 30,
+        }
         if args.custom_network:
-            command += [
-                '--custom-network',
-                args.custom_network
-            ]
-        exit(subprocess.call(command, env=os.environ))
+            eval_args['custom_network'] = args.custom_network
+        main(eval_args)
 
 
 if __name__ == '__main__':
